[Aptos Data Client] Create a new global data summary file.
