+++ state-sync/aptos-data-client/src/aptosnet/poller.rs
@@ -8,7 +8,8 @@ use crate::{
         metrics,
         metrics::{set_gauge, start_request_timer, DataType},
     },
-    AptosDataClientInterface, Error, GlobalDataSummary, Response,
+    global_summary::GlobalDataSummary,
+    AptosDataClientInterface, Error, Response,
 };
 use aptos_config::network_id::PeerNetworkId;
 use aptos_logger::{debug, info, sample, sample::SampleRate, warn};
