+++ state-sync/state-sync-v2/data-streaming-service/src/tests/stream_engine.rs
@@ -9,7 +9,10 @@ use crate::{
     streaming_client::{GetAllEpochEndingLedgerInfosRequest, StreamRequest},
     tests::utils::initialize_logger,
 };
-use aptos_data_client::{GlobalDataSummary, OptimalChunkSizes, ResponsePayload};
+use aptos_data_client::{
+    global_summary::{GlobalDataSummary, OptimalChunkSizes},
+    ResponsePayload,
+};
 use aptos_id_generator::U64IdGenerator;
 use aptos_storage_service_types::responses::CompleteDataRange;
 use claims::{assert_matches, assert_ok};
